hotfix: categoryMeta 외부 클래스로 변경
